Solution

+++ cinema/views.py
@@ -1,32 +1,51 @@
+from rest_framework import generics, mixins, viewsets
 from rest_framework.views import APIView
-from rest_framework.generics import (
-    GenericAPIView,
-    ListCreateAPIView,
-    RetrieveUpdateDestroyAPIView
-)
-from rest_framework.viewsets import GenericViewSet, ModelViewSet
-from rest_framework.mixins import (
-    ListModelMixin,
-    CreateModelMixin,
-    RetrieveModelMixin,
-    UpdateModelMixin,
-    DestroyModelMixin
-)
 from rest_framework.response import Response
 from rest_framework import status
 
-from django.shortcuts import get_object_or_404
-
-from cinema.models import Genre, Actor, CinemaHall, Movie
+from cinema.models import Actor, Genre, CinemaHall, Movie
 from cinema.serializers import (
-    GenreSerializer,
     ActorSerializer,
+    GenreSerializer,
     CinemaHallSerializer,
     MovieSerializer
 )
 
 
-class GenreListCreateAPIView(APIView):
+class ActorList(mixins.ListModelMixin,
+                mixins.CreateModelMixin,
+                generics.GenericAPIView):
+    queryset = Actor.objects.all()
+    serializer_class = ActorSerializer
+
+    def get(self, request, *args, **kwargs):
+        return self.list(request, *args, **kwargs)
+
+    def post(self, request, *args, **kwargs):
+        return self.create(request, *args, **kwargs)
+
+
+class ActorDetail(mixins.RetrieveModelMixin,
+                  mixins.UpdateModelMixin,
+                  mixins.DestroyModelMixin,
+                  generics.GenericAPIView):
+    queryset = Actor.objects.all()
+    serializer_class = ActorSerializer
+
+    def get(self, request, *args, **kwargs):
+        return self.retrieve(request, *args, **kwargs)
+
+    def put(self, request, *args, **kwargs):
+        return self.update(request, *args, **kwargs)
+
+    def patch(self, request, *args, **kwargs):
+        return self.partial_update(request, *args, **kwargs)
+
+    def delete(self, request, *args, **kwargs):
+        return self.destroy(request, *args, **kwargs)
+
+
+class GenreList(APIView):
     def get(self, request):
         genres = Genre.objects.all()
         serializer = GenreSerializer(genres, many=True)
@@ -40,17 +59,24 @@ class GenreListCreateAPIView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-class GenreDetailAPIView(APIView):
+class GenreDetail(APIView):
     def get_object(self, pk):
-        return get_object_or_404(Genre, pk=pk)
+        try:
+            return Genre.objects.get(pk=pk)
+        except Genre.DoesNotExist:
+            return None
 
     def get(self, request, pk):
         genre = self.get_object(pk)
+        if not genre:
+            return Response(status=status.HTTP_404_NOT_FOUND)
         serializer = GenreSerializer(genre)
         return Response(serializer.data)
 
     def put(self, request, pk):
         genre = self.get_object(pk)
+        if not genre:
+            return Response(status=status.HTTP_404_NOT_FOUND)
         serializer = GenreSerializer(genre, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -59,6 +85,8 @@ class GenreDetailAPIView(APIView):
 
     def patch(self, request, pk):
         genre = self.get_object(pk)
+        if not genre:
+            return Response(status=status.HTTP_404_NOT_FOUND)
         serializer = GenreSerializer(genre, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
@@ -67,75 +95,22 @@ class GenreDetailAPIView(APIView):
 
     def delete(self, request, pk):
         genre = self.get_object(pk)
+        if not genre:
+            return Response(status=status.HTTP_404_NOT_FOUND)
         genre.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
-class ActorListCreateView(GenericAPIView):
-    queryset = Actor.objects.all()
-    serializer_class = ActorSerializer
-
-    def get(self, request):
-        actors = self.get_queryset()
-        serializer = self.get_serializer(actors, many=True)
-        return Response(serializer.data)
-
-    def post(self, request):
-        serializer = self.get_serializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-class ActorDetailView(GenericAPIView):
-    queryset = Actor.objects.all()
-    serializer_class = ActorSerializer
-
-    def get_object(self):
-        return get_object_or_404(self.get_queryset(), pk=self.kwargs["pk"])
-
-    def get(self, request, pk):
-        actor = self.get_object()
-        serializer = self.get_serializer(actor)
-        return Response(serializer.data)
-
-    def put(self, request, pk):
-        actor = self.get_object()
-        serializer = self.get_serializer(actor, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-    def patch(self, request, pk):
-        actor = self.get_object()
-        serializer = self.get_serializer(
-            actor,
-            data=request.data,
-            partial=True
-        )
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-    def delete(self, request, pk):
-        actor = self.get_object()
-        actor.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
-
-class CinemaHallViewSet(GenericViewSet,
-                        ListModelMixin,
-                        CreateModelMixin,
-                        RetrieveModelMixin,
-                        UpdateModelMixin,
-                        DestroyModelMixin):
+class CinemaHallViewSet(viewsets.GenericViewSet,
+                        mixins.ListModelMixin,
+                        mixins.CreateModelMixin,
+                        mixins.RetrieveModelMixin,
+                        mixins.UpdateModelMixin,
+                        mixins.DestroyModelMixin):
     queryset = CinemaHall.objects.all()
     serializer_class = CinemaHallSerializer
 
 
-class MovieViewSet(ModelViewSet):
+class MovieViewSet(viewsets.ModelViewSet):
     queryset = Movie.objects.all()
     serializer_class = MovieSerializer
